pyproject.toml
[tool.poetry]
name = "amsskamere"
version = "0.1.0"
description = ""
authors = ["2Ths <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
selenium = "^4.23.0"
webdriver-manager = "^4.0.1"
pyppeteer=">=2.0.0"
pillow=">=2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
